Update: ファイルを移動．登録をLoginと対称的な英語に変更
