+++ app/build.gradle
@@ -65,14 +65,20 @@ dependencies {
     implementation 'com.google.android.material:material:1.4.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.1'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
+
+
+
+    // Jetpack Compose
+    implementation 'androidx.activity:activity-compose:1.5.1'
     implementation "androidx.compose.ui:ui:1.2.1"
     implementation "androidx.compose.material:material:1.2.1"
     implementation "androidx.compose.ui:ui-tooling-preview:1.2.1"
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
-    implementation 'androidx.activity:activity-compose:1.5.1'
     androidTestImplementation "androidx.compose.ui:ui-test-junit4:1.2.1"
     debugImplementation "androidx.compose.ui:ui-tooling:1.2.1"
     debugImplementation "androidx.compose.ui:ui-test-manifest:1.2.1"
+    implementation "androidx.navigation:navigation-compose:2.5.1"
+
     // Navigation
     def nav_version = '2.3.5'
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
