+++ app/src/main/java/com/websarva/wings/android/flat/ui/startup/StartupViewModel.kt
@@ -1,0 +1,10 @@
+package com.websarva.wings.android.flat.ui.startup
+
+import androidx.lifecycle.SavedStateHandle
+import androidx.lifecycle.ViewModel
+
+class StartupViewModel(
+    saveStateHandle: SavedStateHandle
+) : ViewModel() {
+
+}
