+++ app/src/main/java/com/websarva/wings/android/flat/ui/startup/login/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.websarva.wings.android.flat.viewmodel
+package com.websarva.wings.android.flat.ui.startup.login
 
 import android.util.Log
 import androidx.lifecycle.LiveData
