--- app/src/main/java/com/websarva/wings/android/flat/viewmodel/StartupViewModel.kt
@@ -1,10 +1,0 @@
-package com.websarva.wings.android.flat.viewmodel
-
-import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
-
-class StartupViewModel(
-    saveStateHandle: SavedStateHandle
-) : ViewModel() {
-
-}
