+++ app/src/main/res/layout/fragment_account_registration.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".view.AccountRegistrationFragment">
+    tools:context=".ui.startup.signUp.SignUpFragment">
 
     <TextView
         android:id="@+id/tv_account_registration_title"
